requirements.txt
Flask
requests
beautifulsoup4
psycopg2-binary
gunicorn
feedparser